+++ presentation/src/main/java/com/practical/presentation/screens/CharactersListScreen.kt
@@ -28,6 +28,7 @@ import coil.compose.AsyncImage
 import com.practical.domain.CharactersListModel
 import com.practical.presentation.R
 import com.practical.presentation.UiState
+import com.practical.presentation.common.ErrorMessage
 import com.practical.presentation.ui.theme.dimens
 import com.practical.presentation.viewmodel.CharacterViewModel
